move translation to separate file
